requirements.txt
beautifulsoup4==4.9.3
colored==2.2.3
impacket==0.10.0
psycopg2_binary==2.9.6
pwn==1.0
Requests==2.31.0
selenium==4.10.0
selenium_wire==5.1.0
termcolor==2.3.0
